Playground: migrate to custom properties
