+++ packages/cubejs-playground/src/QueryBuilder/ButtonDropdown.js
@@ -2,8 +2,6 @@ import React from 'react';
 import * as PropTypes from 'prop-types';
 import { Dropdown } from 'antd';
 import { Button } from '../components';
-// import styled from 'styled-components';
-// import vars from '../variables';
 
 const ButtonDropdown = ({ overlay, ...buttonProps }) => (
   <Dropdown
@@ -15,13 +13,5 @@ const ButtonDropdown = ({ overlay, ...buttonProps }) => (
 ButtonDropdown.propTypes = {
   overlay: PropTypes.object.isRequired,
 };
-
-// const styledButtonDropdown = styled(ButtonDropdown)`
-//   &.ant-btn {
-//     border-style: dashed;
-//   }
-// `;
-//
-// export default styledButtonDropdown;
 
 export default ButtonDropdown;
