+++ packages/cubejs-playground/src/components/Input.js
@@ -1,17 +1,16 @@
 import styled from 'styled-components';
 import { Input as AntdInput } from 'antd';
-import vars from '../variables';
 
 const StyledInput = styled(AntdInput)`
   && .ant-select-selector {
     padding: 5px 12px;
     height: auto;
-    border-color: 1px solid ${vars.dark05Color};
-    color: ${vars.textColor};
+    border-color: 1px solid var(--dark-05-color);
+    color: var(--text-color);
     
     &:hover, &:active, &:focus {
-      border-color: ${vars.purple04Color};
-      color: ${vars.primaryColor};
+      border-color: var(--purple-04-color);
+      color: var(--primary-color);
     }
   }
 `;
