+++ packages/cubejs-playground/src/components/Pivot/Item.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
 import { DragOutlined } from '@ant-design/icons';
 import { Typography } from 'antd';
-import vars from '../../variables';
 
 export default function Item({ id, index }) {
   return (
@@ -14,7 +13,7 @@ export default function Item({ id, index }) {
           {...dragHandleProps}
           style={{
             ...draggableProps.style,
-            border: `1px dashed ${vars.dark05Color}`,
+            border: `1px dashed var(--dark-05-color)`,
             borderRadius: 4,
             padding: '5px 12px',
             lineHeight: '22px',
