+++ packages/cubejs-playground/src/components/SectionHeader.js
@@ -1,6 +1,5 @@
 // import React from 'react';
 import styled from 'styled-components';
-import vars from '../variables';
 
 const SectionHeader = styled.div`
   font-weight: 500;
@@ -8,7 +7,7 @@ const SectionHeader = styled.div`
   line-height: 18px;
   letter-spacing: 0.04em;
   text-transform: uppercase;
-  color: ${vars.dark03Color};
+  color: var(--dark-03-color);
   margin-bottom: 8px;
 `;
