+++ packages/cubejs-playground/src/components/Select.js
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 import { Select as AntdSelect } from 'antd';
-import vars from '../variables';
 
 const StyledSelect = styled(AntdSelect)`
   &&& {
@@ -32,14 +31,14 @@ const StyledSelect = styled(AntdSelect)`
       
     .ant-select-selector {  
       height: auto;
-      border-color: 1px solid ${vars.dark05Color};
-      color: ${vars.textColor};
+      border-color: 1px solid var(--dark-05-color);
+      color: var(--text-color);
       font-size: 14px;
       line-height: 22px;
       
       &:hover, &:active, &:focus {
-        border-color: ${vars.purple04Color};
-        color: ${vars.primaryColor};
+        border-color: var(--purple-04-color);
+        color: var(--primary-color);
       }    
     }
