+++ packages/cubejs-playground/src/variables.js
@@ -1,29 +1,65 @@
-const varsJSON = require('./raw-variables');
+const colors = {
+  pink: '255, 100, 146',
+  purple: '122, 119, 255',
+  'purple-03': '175, 173, 255',
+  'purple-04': '202, 201, 255',
+  text: '91, 92, 125',
+  'dark-01': '20, 20, 70',
+  'dark-02': '67, 67, 107',
+  'dark-03': '161, 161, 181',
+  'dark-05': '213, 213, 226',
+  light: '243, 243, 251',
+};
 
-function toCamelCase(str) {
-  return str.replace(/-[a-z0-9]/g, s => s.replace('-', '').toUpperCase());
+function color(name, opacity = 1) {
+  return `rgba(${colors[name]}, ${opacity})`;
 }
 
-const vars = {};
+module.exports = {
+  'active-bg': color('purple', .1),
+  'primary-bg': color('purple', .1),
+  'primary-1': color('purple', .9),
+  'primary-2': color('purple', .8),
+  'primary-3': color('purple', .7),
+  'primary-4': color('purple', .6),
+  'primary-5': color('purple', .5),
+  'primary-6': color('purple', .4),
+  'primary-7': color('purple', .3),
+  'primary-8': color('purple', .2),
+  'primary-9': color('purple', .1),
+  'primary-10': color('purple', 0),
+  'light-5': color('light', .5),
+  'remove-btn-bg': color('purple', .1),
+  'remove-btn-hover-bg': color('purple', .2),
+  'primary-color': color('purple'),
 
-// Resolve variable links and create LESS variable map.
-Object.keys(varsJSON)
-  .forEach((key) => {
-    let value = varsJSON[key];
+  'heading-color': color('dark', .65),
+  'link-color': color('purple'),
+  'info-color': color('purple'),
+  'layout-body-background': '#f3f3fc',
+  'layout-header-background': '#eeeef5',
+  'menu-highlight-color': color('dark-01'),
+  'item-hover-bg': color('light'),
+  'layout-header-height': '48px',
 
-    while (value.includes('@')) {
-      const newValue = value.replace(/@([a-z0-9-]+)/gi, (s, s1) => {
-        return varsJSON[s1] || s;
-      });
+  'font-family': '\'DM Sans\', sans-serif',
+  'menu-item-font-size': '15px',
+  'btn-primary-shadow': 'none',
+  'btn-text-shadow': 'none',
+  'modal-body-padding': '32px',
+  'form-item-margin-bottom': '23px',
+  'disabled-color': color('dark-03'),
+  'disabled-bg': color('dark-05', .2),
 
-      if (newValue !== value) {
-        value = newValue;
-      } else {
-        break;
-      }
-    }
+  // menu
+  'menu-item-active-bg': color('light'),
 
-    vars[toCamelCase(key)] = value;
+  'font-size-base': '14px',
+  'border-radius-base': '4px',
+  'padding-lg': '16px',
+};
+
+Object.keys(colors)
+  .forEach((name) => {
+    module.exports[`${name}-color`] = color(name);
   });
-
-module.exports = vars;
